+++ dioptas/tests/controller_tests/test_BatchController_part2.py
@@ -18,36 +18,6 @@ unittest_data_path = os.path.join(os.path.dirname(__file__), '../data')
 jcpds_path = os.path.join(unittest_data_path, 'jcpds')
 
 
-class TestMouseEvent:
-    def __init__(self, key=None, diff=None):
-        self.key_value = key
-        self.diff = diff
-
-        class TestCoord:
-            def x(self):
-                return 100
-
-            def y(self):
-                return 100
-
-        self.coord = TestCoord()
-
-    def key(self):
-        return self.key_value
-
-    def x(self):
-        return self.diff
-
-    def angleDelta(self):
-        return self.coord
-
-    def modifiers(self):
-        return QtCore.Qt.CoverWindow
-
-    def button(self):
-        return QtCore.Qt.CoverWindow
-
-
 class BatchControllerTest(QtTest):
     def setUp(self):
         self.working_dir = {'image': ''}
